--- frontend/src/app/core/models/user.ts
@@ -1,5 +1,0 @@
-export interface User {
-  id: number;
-  username: string;
-  email: string;
-}
