+++ frontend/src/app/features/map/services/models/geocode-dto.ts
@@ -1,0 +1,5 @@
+export interface GeocodeDTO {
+    label: string;
+    lat: number;
+    lng: number;
+}
