+++ frontend/src/app/features/map/services/models/lat-lng.ts
