+++ frontend/src/app/features/map/services/models/ors-directions-response.ts
@@ -1,0 +1,5 @@
+import { OrsFeature } from "./ors-feature";
+
+export interface OrsDirectionsResponse {
+    features: OrsFeature[];
+}
