+++ frontend/src/app/features/map/services/models/ors-feature.ts
@@ -1,0 +1,7 @@
+import { OrsGeometry } from "./ors-geometry";
+import { OrsSummary } from "./ors-summary";
+
+export interface OrsFeature {
+    geometry: OrsGeometry;
+    properties: { summary: OrsSummary };
+}
