+++ frontend/src/app/features/map/services/models/ors-geometry.ts
@@ -1,0 +1,3 @@
+export interface OrsGeometry {
+    coordinates: [number, number][]; // [lng, lat]
+}
