+++ frontend/src/app/features/map/services/models/ors-summary.ts
@@ -1,0 +1,4 @@
+export interface OrsSummary {
+    distance: number; // metres
+    duration: number; // seconds
+}
