+++ frontend/src/app/features/tours/models/log/create-tour-log-form.ts
@@ -1,0 +1,8 @@
+export interface CreateTourLogForm {
+    tour_id: number;
+    comment: string;
+    difficulty: number;
+    rating: number;
+    total_time_m: number;
+    total_distance_km: number;
+}
