+++ frontend/src/app/features/tours/models/log/create-tour-log-payload.ts
@@ -1,0 +1,8 @@
+export interface CreateTourLogPayload {
+    tour_id: number;
+    comment: string;
+    difficulty: number;
+    rating: number;
+    total_time_s: number;
+    total_distance_km: number;
+}
