+++ frontend/src/app/features/tours/models/log/update-tour-log-form.ts
@@ -1,0 +1,7 @@
+export interface UpdateTourLogForm {
+    comment?: string;
+    difficulty?: number;
+    rating?: number;
+    total_time_m?: number;
+    total_distance_km?: number;
+}
