+++ frontend/src/app/features/tours/models/log/update-tour-log-payload.ts
@@ -1,0 +1,7 @@
+export interface UpdateTourLogPayload {
+    comment?: string;
+    difficulty?: number;
+    rating?: number;
+    total_time_s?: number;
+    total_distance_km?: number;
+}
