+++ frontend/src/app/features/tours/models/tour/create-tour-payload.ts
@@ -1,0 +1,9 @@
+import { GeocodeDTO } from "../../../map/services/models/geocode-dto";
+
+export interface CreateTourPayload {
+    name: string;
+    description?: string;
+    fromGeocode: GeocodeDTO;
+    toGeocode: GeocodeDTO;
+    profile: string;
+}
