+++ frontend/src/app/features/tours/models/tour/tour-summary.ts
