+++ frontend/src/app/features/tours/models/tour/update-tour-payload.ts
@@ -1,0 +1,9 @@
+import { GeocodeDTO } from "../../../map/services/models/geocode-dto";
+
+export interface UpdateTourPayload {
+    name: string;
+    description?: string;
+    fromGeocode: GeocodeDTO;
+    toGeocode: GeocodeDTO;
+    profile: string;
+}
