+++ frontend/src/app/features/user/components/login/login.ts
@@ -3,6 +3,7 @@ import { LoginForm } from './models';
 import { email, form, FormField, required } from '@angular/forms/signals';
 import { UserService } from '../../services/user.service';
 import { Router } from '@angular/router';
+import { LoginRequest } from '../../models/login-request';
 
 @Component({
   selector: 'app-login',
@@ -30,11 +31,18 @@ export class Login {
   onSubmit(event: Event) {
     event.preventDefault();
 
-    if (this.loginForm.email().invalid() || this.loginForm.password().invalid()) return;
+    // validate user input
+    if (this.loginForm.email().invalid() || this.loginForm.password().invalid()) {
+      return;
+    }
 
-    const { email, password } = this.loginModel();
+    // create request object
+    const request: LoginRequest = {
+      email: this.loginModel().email,
+      password: this.loginModel().password
+    };
 
-    this.userService.login(email, password).subscribe({
+    this.userService.login(request).subscribe({
       next: () => this.router.navigate(['/profile']),
       error: (err: Error) => alert(err.message),
     });
