+++ frontend/src/app/features/user/models/auth-response.ts
@@ -1,0 +1,3 @@
+export interface AuthResponse {
+    token: string;
+}
