+++ frontend/src/app/features/user/models/login-request.ts
@@ -1,0 +1,4 @@
+export interface LoginRequest {
+    email: string;
+    password: string;
+}
