+++ frontend/src/app/features/user/models/register-request.ts
@@ -1,0 +1,5 @@
+export interface RegisterRequest {
+    email: string;
+    password: string;
+    passwordRepeat: string;
+}
